+++ src/hankel_transform.c
@@ -44,7 +44,7 @@ int validate_f_max(strategy_params strategy_params) {
  * @param strategy_params  struct containing params for the specific strategy
  *                         (please use the docs to find params required by each strategy)
  */ 
-double hankel_transform(
+int hankel_transform(
     int nu, 
     double (*f)(double, double (*)[50]), 
     const double *x,
